set up testing env for firebase config
